refactor: intern form cleanup, create camera component
